fixed comparison of different type

+++ ft_insertion_sort.c
@@ -6,7 +6,7 @@
 /*   By: lmeubrin <lmeubrin@student.42berlin.d      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/01 13:50:40 by lmeubrin          #+#    #+#             */
-/*   Updated: 2025/04/08 11:54:46 by lmeubrin         ###   ########.fr       */
+/*   Updated: 2025/04/23 14:22:47 by lmeubrin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ void	ft_insertion_sort(float *arr, size_t size)
 	float	key;
 
 	i = 1;
-	while (i < size)
+	while (i < (int) size)
 	{
 		key = arr[i];
 		j = i - 1;
